fix: quality of life updates & container existence checks

+++ cmd/delete.go
@@ -12,23 +12,26 @@ import (
 func deleteContainer(containerName string) {
 	projectPath, err := internal.ContainerProjectPath(containerName)
 	if err != nil {
-		fmt.Printf("[!] failed to resolve project path for %s: %v\n", containerName, err)
+		fmt.Printf("[!] failed to resolve project path for '%s': %v\n", containerName, err)
 	} else {
 		sdDir := filepath.Join(projectPath, ".stormdrain")
 		if err = os.RemoveAll(sdDir); err != nil {
-			fmt.Printf("[!] failed to remove data directory %s: %v\n", sdDir, err)
+			fmt.Printf("[!] failed to remove data directory '%s': %v\n", sdDir, err)
 		}
 	}
 
+	fmt.Printf("[~] removing container '%s'... ", containerName)
 	if err := internal.PodmanRemove(containerName); err != nil {
-		fmt.Printf("[!] failed to remove container %s: %v\n", containerName, err)
+		fmt.Println("failed")
+		fmt.Printf("[!] failed to remove container '%s': %v\n", containerName, err)
 	}
+	fmt.Println("done")
 }
 
 func CmdDelete(args []string) {
 	fs := flag.NewFlagSet("rm", flag.ExitOnError)
 	fs.Usage = func() {
-		fmt.Printf("usage: %s rm [name]\n", os.Args[0])
+		fmt.Println("[?] usage: stormdrain rm [name]")
 	}
 	fs.Parse(args)
 
@@ -38,10 +41,15 @@ func CmdDelete(args []string) {
 	} else {
 		cwd, err := os.Getwd()
 		if err != nil {
-			fmt.Printf("[!] failed resolving cwd: %v\n", err)
+			fmt.Printf("[!] failed to resolve cwd: %v\n", err)
 			os.Exit(1)
 		}
 		containerName = filepath.Base(cwd)
+	}
+
+	if !internal.ContainerExists(containerName) {
+		fmt.Printf("[!] container '%s' does not exist\n", containerName)
+		os.Exit(1)
 	}
 
 	deleteContainer(containerName)
@@ -50,15 +58,16 @@ func CmdDelete(args []string) {
 func CmdDeleteAll() {
 	// NOTE: instead of utilizing podman's built-in batch removal system, we opt for cleaner design here
 	//		 (as the decrease in performance is quite substantial with lower amounts of containers to remove)
-	ids, err := internal.StormdrainContainerIDs()
+	ids, err := internal.ListContainerIDs()
 	if err != nil {
 		fmt.Printf("[!] failed to list stormdrain containers: %v\n", err)
 		os.Exit(1)
 	}
 	if len(ids) == 0 {
-		fmt.Println("no stormdrain containers found")
+		fmt.Println("[+] no stormdrain containers found")
 		return
 	}
+	fmt.Printf("[~] removing a total of %d containers\n", len(ids))
 
 	for _, id := range ids {
 		deleteContainer(id)
